bt-235: * Fix badge component file structure
