--- frontend/src/bundles/common/components/badge/components.ts
@@ -1,1 +1,0 @@
-export { BadgeIcon } from './components/badge-icon/badge-icon.js';
