+++ frontend/src/bundles/common/components/badge/components/badge-icon.tsx
